added more items, ui refactoring

+++ app/src/main/java/org/sugarandrose/app/data/model/LocalPost.kt
@@ -15,7 +15,7 @@ import paperparcel.PaperParcelable
  */
 
 @PaperParcel
-open class LocalPost(@PrimaryKey override var id: Int,
+open class LocalPost(@PrimaryKey override var id: Long,
                      override var date: String,
                      var title: String,
                      var url: String,
